AST: parseMemberExpression() supports secondary expression
It's possible for a member expression to be a secondary expression to
a different AstNode. For example, "console.log(functional_call());". Here,
the member expression 'console.log' should be the 'identifier' of the call
expression to 'log'.

This patch also requires that the right hand side of a MemberExpression is an
identifier. It's not valid syntax for it to be anything else, so we make that
guarantee from the start.