Tokenizer: Continue parsing the current token after scratch space was handled
Previously after handling the scratch space the current token could be skipped if
the scratch space contained one of the following: a numeric literal, a string litreal,
a boolean litreal, or a null litreal. These changes update the behavior to continue parsing
the current token after handling the scratch space.