Tokenizer: Append to scratch if a punctuator is part of a string
Previously when a punctuator occurred during a string, the string literal
would be handled as an identifier so that the next punctuator token could be
handled as a punctuator. Now, if the scratch space contains the start of a
string literal and a punctuator is encountered, it is added as a character
in the string literal.