LibTest: Initial commit of testing framework
This commit changes a bunch of testing files and adds a couple of new
and shiny classes. The important classes are Test.java, Suite.java, and
TestRunner.java. Test and Suite are annotations that define which
methods and from which classes (respectively) to run test with. The
TestRunner will automatically search the classpath for classes that have
a @Suite annotation and run all methods in that class that have the
@Test annotation. Adding future tests should be as simple as adding the
apporiate annotations in the correct spots :).